+++ test/uk/gov/hmrc/trusts/utils/DataExamples.scala
@@ -128,11 +128,13 @@ trait DataExamples extends  JsonRequests {
 
   def trustWithValues(indBenficiaryDob : String ="2001-01-01",
                       settlorNino :String = "ST019092",
-                      settlorDob :String = "2001-01-01"
+                      settlorDob :String = "2001-01-01",
+                      settlorUtr :String = "1234561235"
                     ) : String = {
     val json = getJsonValueFromFile("trusts-dynamic.json")
     json.toString().replace("{indBeneficiaryDob}", indBenficiaryDob)
       .replace("{settlorNino}", settlorNino).replace("{settlorDob}", settlorDob)
+      .replace("{settlorUtr}", settlorUtr)
   }
 
   def willTrustWithValues(
